+++ internal/transfer/provider_mbox_test.go
@@ -35,25 +35,28 @@ func newTestMBOXProvider(path string) *MBOXProvider {
 }
 
 func TestMBOXProviderMailboxes(t *testing.T) {
-	provider := newTestMBOXProvider("")
-
 	tests := []struct {
+		provider      *MBOXProvider
 		includeEmpty  bool
 		wantMailboxes []Mailbox
 	}{
-		{true, []Mailbox{
+		{newTestMBOXProvider(""), true, []Mailbox{
+			{Name: "All Mail"},
 			{Name: "Foo"},
+			{Name: "Bar"},
 			{Name: "Inbox"},
 		}},
-		{false, []Mailbox{
+		{newTestMBOXProvider(""), false, []Mailbox{
+			{Name: "All Mail"},
 			{Name: "Foo"},
+			{Name: "Bar"},
 			{Name: "Inbox"},
 		}},
 	}
 	for _, tc := range tests {
 		tc := tc
 		t.Run(fmt.Sprintf("%v", tc.includeEmpty), func(t *testing.T) {
-			mailboxes, err := provider.Mailboxes(tc.includeEmpty, false)
+			mailboxes, err := tc.provider.Mailboxes(tc.includeEmpty, false)
 			r.NoError(t, err)
 			r.Equal(t, tc.wantMailboxes, mailboxes)
 		})
@@ -67,14 +70,26 @@ func TestMBOXProviderTransferTo(t *testing.T) {
 	defer rulesClose()
 	setupMBOXRules(rules)
 
-	testTransferTo(t, rules, provider, []string{
+	msgs := testTransferTo(t, rules, provider, []string{
+		"All Mail.mbox:1",
+		"All Mail.mbox:2",
 		"Foo.mbox:1",
 		"Inbox.mbox:1",
 	})
+	got := map[string][]string{}
+	for _, msg := range msgs {
+		got[msg.ID] = msg.targetNames()
+	}
+	r.Equal(t, map[string][]string{
+		"All Mail.mbox:1": {"Archive", "Foo"}, // Bar is not in rules.
+		"All Mail.mbox:2": {"Archive", "Foo"},
+		"Foo.mbox:1":      {"Foo"},
+		"Inbox.mbox:1":    {"Inbox"},
+	}, got)
 }
 
 func TestMBOXProviderTransferFrom(t *testing.T) {
-	dir, err := ioutil.TempDir("", "eml")
+	dir, err := ioutil.TempDir("", "mbox")
 	r.NoError(t, err)
 	defer os.RemoveAll(dir) //nolint[errcheck]
 
@@ -94,7 +109,7 @@ func TestMBOXProviderTransferFrom(t *testing.T) {
 }
 
 func TestMBOXProviderTransferFromTo(t *testing.T) {
-	dir, err := ioutil.TempDir("", "eml")
+	dir, err := ioutil.TempDir("", "mbox")
 	r.NoError(t, err)
 	defer os.RemoveAll(dir) //nolint[errcheck]
 
@@ -103,17 +118,51 @@ func TestMBOXProviderTransferFromTo(t *testing.T) {
 
 	rules, rulesClose := newTestRules(t)
 	defer rulesClose()
-	setupEMLRules(rules)
+	setupMBOXRules(rules)
 
 	testTransferFromTo(t, rules, source, target, 5*time.Second)
 
 	checkMBOXFileStructure(t, dir, []string{
+		"Archive.mbox",
 		"Foo.mbox",
 		"Inbox.mbox",
 	})
 }
 
+func TestMBOXProviderGetMessageTargetsReturnsOnlyOneFolder(t *testing.T) {
+	provider := newTestMBOXProvider("")
+
+	folderA := Mailbox{Name: "Folder A", IsExclusive: true}
+	folderB := Mailbox{Name: "Folder B", IsExclusive: true}
+	labelA := Mailbox{Name: "Label A", IsExclusive: false}
+	labelB := Mailbox{Name: "Label B", IsExclusive: false}
+	labelC := Mailbox{Name: "Label C", IsExclusive: false}
+
+	rule1 := &Rule{TargetMailboxes: []Mailbox{folderA, labelA, labelB}}
+	rule2 := &Rule{TargetMailboxes: []Mailbox{folderB, labelC}}
+	rule3 := &Rule{TargetMailboxes: []Mailbox{folderB}}
+
+	tests := []struct {
+		rules         []*Rule
+		wantMailboxes []Mailbox
+	}{
+		{[]*Rule{}, []Mailbox{}},
+		{[]*Rule{rule1}, []Mailbox{folderA, labelA, labelB}},
+		{[]*Rule{rule1, rule2}, []Mailbox{folderA, labelA, labelB, labelC}},
+		{[]*Rule{rule1, rule3}, []Mailbox{folderA, labelA, labelB}},
+		{[]*Rule{rule3, rule1}, []Mailbox{folderB, labelA, labelB}},
+	}
+	for _, tc := range tests {
+		tc := tc
+		t.Run(fmt.Sprintf("%v", tc.rules), func(t *testing.T) {
+			mailboxes := provider.getMessageTargets(tc.rules, "", []byte(""))
+			r.Equal(t, tc.wantMailboxes, mailboxes)
+		})
+	}
+}
+
 func setupMBOXRules(rules transferRules) {
+	_ = rules.setRule(Mailbox{Name: "All Mail"}, []Mailbox{{Name: "Archive"}}, 0, 0)
 	_ = rules.setRule(Mailbox{Name: "Inbox"}, []Mailbox{{Name: "Inbox"}}, 0, 0)
 	_ = rules.setRule(Mailbox{Name: "Foo"}, []Mailbox{{Name: "Foo"}}, 0, 0)
 }